// app/_modules/book/book.types.ts
export type BookType = {
  id: string;
  title: string;
  authors: string;
  isbn: string;
  language: string;
  category: string;
  description: string;
  textSnippet: string;
  thumbnail: string;
  imgLarge: string;
};

export type BookReferenceType = {
  id: string;
  title: string;
  thumbnail: string;
};